All changes are in response to the creation of a project dedicated to code quality assurance. Much of the Ant task machinery has been moved to that project so that it does not need to be repeated in each of our projects.
Changes made at this time:
Significant updates to build.xml to allow the new QA project to be checked out so that unit tests and things like FindBugs can still be run.

Addition of the svnant libraries in ant/lib so that the above checkout can occur

Removal of lots of resources from lib-test because they are now in the codeqa project.